Refactor compile-heavy runtime units
