+++ src/config_types.hpp
@@ -1,0 +1,550 @@
+/**
+ * @file config_types.hpp
+ * @brief JSON-independent runtime configuration types and defaults.
+ *
+ * This layer owns durable configuration values and their serialized
+ * representation. Transient runtime requests such as `--test-tone` do not
+ * live here. Frequency entries may include optional `@GPIO[H|L]` metadata
+ * that overrides the selected band GPIO for one scheduler slot.
+ *
+ * This project is is licensed under the MIT License. See LICENSE.md
+ * for more information.
+ *
+ * Copyright © 2023-2026 Lee C. Bussy (@LBussy). All rights reserved.
+ *
+ * Permission is hereby granted, free of charge, to any person obtaining a copy
+ * of this software and associated documentation files (the "Software"), to
+ * deal in the Software without restriction, including without limitation the
+ * rights to use, copy, modify, merge, publish, distribute, sublicense, and/or
+ * sell copies of the Software, and to permit persons to whom the Software is
+ * furnished to do so, subject to the following conditions:
+ *
+ * The above copyright notice and this permission notice shall be included in
+ * all copies or substantial portions of the Software.
+ *
+ * THE SOFTWARE IS PROVIDED "AS IS", WITHOUT WARRANTY OF ANY KIND, EXPRESS OR
+ * IMPLIED, INCLUDING BUT NOT LIMITED TO THE WARRANTIES OF MERCHANTABILITY,
+ * FITNESS FOR A PARTICULAR PURPOSE AND NONINFRINGEMENT. IN NO EVENT SHALL THE
+ * AUTHORS OR COPYRIGHT HOLDERS BE LIABLE FOR ANY CLAIM, DAMAGES OR OTHER
+ * LIABILITY, WHETHER IN AN ACTION OF CONTRACT, TORT OR OTHERWISE, ARISING FROM,
+ * OUT OF OR IN CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN THE
+ * SOFTWARE.
+ */
+
+#ifndef CONFIG_TYPES_HPP
+#define CONFIG_TYPES_HPP
+
+#include "band_gpio.hpp"
+#include "band_lookup.hpp"
+#include "wspr_ref_plan.hpp"
+#include "websocket_listener_config.hpp"
+
+#include <array>
+#include <atomic>
+#include <string>
+#include <vector>
+
+inline constexpr int kTransmitGpioUnset = -1;
+inline constexpr int kDefaultTransmitGpio = 4;
+inline constexpr int kDefaultRp1GpioDriveMa = 2;
+inline constexpr std::array<int, 2> kSupportedTransmitGpio = {4, 20};
+inline constexpr int kSelectorGpioUnset = -1;
+inline constexpr double WSPR_AUDIO_OFFSET_HZ = 1500.0;
+inline constexpr int kDefaultSi5351I2cBus = 1;
+inline constexpr int kDefaultSi5351I2cAddress = 0x60;
+inline constexpr int kDefaultSi5351ReferenceHz = 27000000;
+inline constexpr int kDefaultSi5351TxOutput = 0;
+inline constexpr const char *kDefaultSi5351ReferenceSource = "external_tcxo";
+inline constexpr int kDefaultSi5351CrystalLoadCapacitancePf = 10;
+inline constexpr double kDefaultDfcwIntraElementGap = 0.333333;
+inline constexpr double kDefaultDfcwInterCharacterGap = 1.0;
+inline constexpr double kDefaultDfcwInterWordGap = 3.0;
+
+inline constexpr bool is_supported_transmit_gpio(int gpio) noexcept
+{
+    for (int supported_gpio : kSupportedTransmitGpio)
+    {
+        if (gpio == supported_gpio)
+        {
+            return true;
+        }
+    }
+
+    return false;
+}
+
+inline constexpr bool is_valid_selector_gpio(int gpio) noexcept
+{
+    return gpio >= 0 && gpio <= 27;
+}
+
+inline constexpr bool is_supported_rp1_gpio_drive_ma(int drive_ma) noexcept
+{
+    return drive_ma == 2 || drive_ma == 4 || drive_ma == 8 || drive_ma == 12;
+}
+
+struct WsprFrequencyEntry
+{
+    std::string token; ///< Original frequency token without `@GPIO[H|L]`.
+    double dial_frequency_hz = 0.0; ///< Resolved WSPR dial frequency in Hz.
+    int selector_gpio = kSelectorGpioUnset; ///< Optional per-entry selector GPIO.
+    bool selector_gpio_active_high = false; ///< Optional selector GPIO polarity; false means active low.
+    bool allow_band_gpio_fallback = false; ///< True when [Band GPIO] may supply the selector.
+};
+
+/**
+ * @enum ModeType
+ * @brief Specifies the mode of operation for the application.
+ *
+ * This enumeration defines the available modes for operation.
+ * - `WSPR`: Represents the WSPR (Weak Signal Propagation Reporter) transmission mode.
+ * - `TONE`: Represents transient direct-tone runtime behavior.
+ */
+enum class ModeType
+{
+    WSPR, ///< WSPR transmission mode
+    TONE, ///< Test tone generation mode
+    QRSS, ///< Temporary QRSS test mode
+    FSKCW, ///< Temporary FSKCW test mode
+    DFCW ///< Temporary DFCW test mode
+};
+
+enum class WsprPlannerPreference
+{
+    Auto = 0,
+    PreferPaired,
+    RequirePaired
+};
+
+enum class TransmitBackendKind
+{
+    GPIO = 0,
+    RP1_GPCLK,
+    SI5351,
+    SIMULATED
+};
+
+inline constexpr bool transmit_backend_uses_gpio_output(
+    TransmitBackendKind backend) noexcept
+{
+    return backend == TransmitBackendKind::GPIO ||
+        backend == TransmitBackendKind::RP1_GPCLK;
+}
+
+enum class EnableOnBootBehavior
+{
+    Never = 0,
+    Follow,
+    Always
+};
+
+inline constexpr const char *transmit_backend_kind_to_string(
+    TransmitBackendKind backend) noexcept
+{
+    switch (backend)
+    {
+    case TransmitBackendKind::GPIO:
+        return "gpio";
+    case TransmitBackendKind::RP1_GPCLK:
+        return "rp1-gpclk";
+    case TransmitBackendKind::SI5351:
+        return "si5351";
+    case TransmitBackendKind::SIMULATED:
+        return "simulated";
+    }
+
+    return "gpio";
+}
+
+bool transmit_backend_is_compiled(TransmitBackendKind backend) noexcept;
+bool build_has_physical_gpio_capability() noexcept;
+bool transmit_backend_requires_root(TransmitBackendKind backend) noexcept;
+std::string transmit_backend_unavailable_message(TransmitBackendKind backend);
+
+inline constexpr const char *enable_on_boot_behavior_to_string(
+    EnableOnBootBehavior behavior) noexcept
+{
+    switch (behavior)
+    {
+    case EnableOnBootBehavior::Never:
+        return "Never";
+    case EnableOnBootBehavior::Follow:
+        return "Follow";
+    case EnableOnBootBehavior::Always:
+        return "Always";
+    }
+
+    return "Never";
+}
+
+struct WsprModeConfig
+{
+    std::string callsign;
+    std::string grid_square;
+    int power_dbm = 0;
+    std::string frequencies;
+    std::string frequency_profile = "existing_common";
+    WsprBandPreferences band_preferences;
+    double audio_offset_hz = WSPR_AUDIO_OFFSET_HZ;
+    WsprPlannerPreference planner_preference = WsprPlannerPreference::Auto;
+};
+
+struct QrssModeConfig
+{
+    std::string message;
+    double frequency_hz = 0.0;
+    double dot_seconds = 0.0;
+};
+
+struct FskcwModeConfig
+{
+    std::string message;
+    double mark_frequency_hz = 0.0;
+    double space_frequency_hz = 0.0;
+    double dot_seconds = 0.0;
+};
+
+struct DfcwModeConfig
+{
+    std::string message;
+    double dot_frequency_hz = 0.0;
+    double dash_frequency_hz = 0.0;
+    double dot_seconds = 0.0;
+};
+
+inline constexpr const char *wspr_planner_preference_to_string(
+    WsprPlannerPreference preference) noexcept
+{
+    switch (preference)
+    {
+    case WsprPlannerPreference::Auto:
+        return "auto";
+    case WsprPlannerPreference::PreferPaired:
+        return "prefer_paired";
+    case WsprPlannerPreference::RequirePaired:
+        return "require_paired";
+    }
+
+    return "auto";
+}
+
+inline constexpr wspr::TransmissionPlanPreference
+wspr_planner_preference_to_plan_preference(
+    WsprPlannerPreference preference) noexcept
+{
+    switch (preference)
+    {
+    case WsprPlannerPreference::Auto:
+        return wspr::TransmissionPlanPreference::Auto;
+    case WsprPlannerPreference::PreferPaired:
+        return wspr::TransmissionPlanPreference::PreferPaired;
+    case WsprPlannerPreference::RequirePaired:
+        return wspr::TransmissionPlanPreference::RequirePaired;
+    }
+
+    return wspr::TransmissionPlanPreference::Auto;
+}
+
+/**
+ * @brief Global configuration instance for argument parsing and runtime settings.
+ *
+ * @details
+ * Holds all command-line and runtime configuration data not managed directly
+ * by the INI file system. Initialized globally and used throughout the application.
+ *
+ * @see ArgParserConfig, ini, iniMonitor
+ */
+struct ArgParserConfig
+{
+    // Runtime
+    bool transmit; ///< Transmission mode enabled.
+    EnableOnBootBehavior enable_on_boot; ///< Persisted transmit behavior after reboot.
+
+    // WSPR
+    std::string callsign;    ///< WSPR callsign.
+    std::string grid_square; ///< 4- or 6-character Maidenhead locator.
+    int power_dbm;           ///< Transmit power in dBm.
+    std::string frequencies; ///< Space-separated user-facing WSPR dial frequency list.
+
+    // Resolved active backend values
+    int tx_pin; ///< Active GPIO pin number for RF transmit control.
+
+    // Runtime
+    double ppm;      ///< Active backend PPM frequency correction.
+    bool use_system_clock_frequency_estimate; ///< Active GPIO provider-estimate setting.
+    bool use_offset; ///< Enable WSPR random frequency offset.
+    int power_level; ///< Active backend RF power level.
+    TransmitBackendKind transmit_backend; ///< RF hardware backend.
+    int gpio_tx_pin; ///< GPIO backend transmit pin.
+    int gpio_power_level; ///< GPIO backend power level (0-7).
+    int rp1_gpio_drive_ma; ///< RP1 GPIO pad drive selection in mA.
+    bool gpio_use_system_clock_frequency_estimate; ///< Enable the GPIO system-clock estimate.
+    double gpio_frequency_residual_ppm; ///< RF residual added to a usable provider estimate.
+    double gpio_manual_ppm; ///< Fixed GPIO correction and final configured fallback.
+    double si5351_ppm; ///< Persisted Si5351 reference-oscillator correction.
+    int si5351_i2c_bus; ///< Si5351 I2C bus number.
+    int si5351_i2c_address; ///< Si5351 I2C slave address.
+    int si5351_reference_hz; ///< Si5351 reference frequency in Hz.
+    std::string si5351_reference_source; ///< external_tcxo or crystal.
+    int si5351_crystal_load_capacitance_pf; ///< Crystal load capacitance in pF.
+    int si5351_tx_output; ///< Si5351 output clock index (0=CLK0).
+    int si5351_power_level; ///< Si5351 drive-strength level (1-4).
+    bool use_led;    ///< Enable TX LED indicator.
+    int led_pin;     ///< GPIO pin for LED indicator.
+    bool use_amp;    ///< Enable external amplifier control GPIO.
+    int amp_pin;     ///< Optional GPIO pin for external amplifier control (-1 = disabled).
+    bool amp_pin_active_high; ///< External amplifier GPIO polarity.
+
+    // Runtime
+    bool enable_web;   ///< Enable the HTTP web UI and WebSocket server.
+    bool enable_http;  ///< Enable the HTTP web UI independently of WebSocket control.
+    int web_port;      ///< Web server port number.
+    int socket_port;   ///< Socket server port number.
+    bool socket_loopback_only; ///< Restrict WebSocket control to loopback.
+    WebSocketLoopbackFamily socket_loopback_family; ///< Command-line-only loopback family policy.
+    bool use_shutdown; ///< Enable GPIO-based shutdown feature.
+    int shutdown_pin;  ///< GPIO pin used to signal shutdown.
+
+    // Command line only
+    bool use_journald;              ///< Route logs to journald instead of streams.
+    bool date_time_log;             ///< Prefix logs with timestamp.
+    bool debug_logging;             ///< Enable DEBUG-level application logging.
+    bool allow_unqualified_frequency; ///< Permit unqualified backend/mode combinations.
+    bool allow_non_amateur_frequency; ///< Permit frequencies outside recognized amateur allocations when combined with the unqualified override.
+    std::string rp1_development_confirmation_json; ///< One direct-CLI RP1 operation confirmation.
+    WsprPlannerPreference wspr_planner_preference; ///< Preferred planner behavior for Type 2/3 pairing.
+    bool loop_tx;                   ///< Repeat transmission cycle.
+    std::atomic<int> tx_iterations; ///< Number of transmission iterations (0 = infinite).
+    double wspr_audio_offset_hz;    ///< Runtime WSPR audio offset constant.
+    double modulation_dot_seconds;  ///< Shared CW dot length.
+    double modulation_fsk_offset_hz; ///< Shared CW shift in Hz.
+    double cw_intra_element_gap; ///< CW intra-element gap in dot-length multiples.
+    double cw_inter_character_gap; ///< CW inter-character gap in dot-length multiples.
+    double cw_inter_word_gap; ///< CW inter-word gap in dot-length multiples.
+    double dfcw_intra_element_gap; ///< DFCW intra-element gap in dot-length multiples.
+    double dfcw_inter_character_gap; ///< DFCW inter-character gap in dot-length multiples.
+    double dfcw_inter_word_gap; ///< DFCW inter-word gap in dot-length multiples.
+    std::string cw_fade_shape; ///< CW envelope fade shape.
+    int cw_fade_in_ms; ///< CW envelope fade-in duration in milliseconds.
+    int cw_fade_out_ms; ///< CW envelope fade-out duration in milliseconds.
+    int cw_fade_slice_ms; ///< CW fade approximation slice duration in milliseconds.
+    int schedule_start_minute;      ///< CW schedule minute offset within the hour.
+    int schedule_start_second;      ///< CW schedule second offset within the minute.
+    int schedule_repeat_minutes;    ///< CW schedule repeat interval in minutes.
+
+    // Runtime variables
+    ModeType mode;                       ///< Current operating mode.
+    WsprModeConfig wspr;                 ///< Long-term WSPR mode configuration.
+    QrssModeConfig qrss;                 ///< Long-term QRSS mode configuration.
+    FskcwModeConfig fskcw;               ///< Long-term FSKCW mode configuration.
+    DfcwModeConfig dfcw;                 ///< Long-term DFCW mode configuration.
+    bool use_ini;                        ///< Load configuration from INI file.
+    std::string ini_filename;            ///< INI file name and path.
+    std::vector<double> wspr_dial_freq_set; ///< Parsed WSPR dial frequencies.
+    std::vector<WsprFrequencyEntry>
+        wspr_frequency_entries; ///< Parsed entries with optional GPIO/polarity metadata.
+    bool frequency_estimate_good;                       ///< A more qualitative measurement of NTP vs simply running
+    std::array<BandGPIOConfig, HAM_BAND_COUNT> band_gpio; ///< Per-band GPIO assignment.
+
+    /**
+     * @brief Default constructor initializing all configuration parameters.
+     */
+    ArgParserConfig()
+        : transmit(true),
+          enable_on_boot(EnableOnBootBehavior::Never),
+          callsign(""),
+          grid_square(""),
+          power_dbm(0),
+          frequencies(""),
+          tx_pin(kTransmitGpioUnset),
+          ppm(0.0),
+          use_system_clock_frequency_estimate(false),
+          use_offset(false),
+          power_level(7),
+          transmit_backend(TransmitBackendKind::GPIO),
+          gpio_tx_pin(kTransmitGpioUnset),
+          gpio_power_level(7),
+          rp1_gpio_drive_ma(kDefaultRp1GpioDriveMa),
+          gpio_use_system_clock_frequency_estimate(false),
+          gpio_frequency_residual_ppm(0.0),
+          gpio_manual_ppm(0.0),
+          si5351_ppm(0.0),
+          si5351_i2c_bus(kDefaultSi5351I2cBus),
+          si5351_i2c_address(kDefaultSi5351I2cAddress),
+          si5351_reference_hz(kDefaultSi5351ReferenceHz),
+          si5351_reference_source(kDefaultSi5351ReferenceSource),
+          si5351_crystal_load_capacitance_pf(kDefaultSi5351CrystalLoadCapacitancePf),
+          si5351_tx_output(kDefaultSi5351TxOutput),
+          si5351_power_level(1),
+          use_led(false),
+          led_pin(-1),
+          use_amp(false),
+          amp_pin(-1),
+          amp_pin_active_high(false),
+          enable_web(true),
+          enable_http(true),
+          web_port(-1),
+          socket_port(-1),
+          socket_loopback_only(false),
+          socket_loopback_family(WebSocketLoopbackFamily::Auto),
+          use_shutdown(false),
+          shutdown_pin(-1),
+          use_journald(false),
+          date_time_log(false),
+          debug_logging(false),
+          allow_unqualified_frequency(false),
+          allow_non_amateur_frequency(false),
+          rp1_development_confirmation_json(""),
+          wspr_planner_preference(WsprPlannerPreference::Auto),
+          loop_tx(false),
+          tx_iterations(0),
+          wspr_audio_offset_hz(WSPR_AUDIO_OFFSET_HZ),
+          modulation_dot_seconds(3.0),
+          modulation_fsk_offset_hz(500.0),
+          cw_intra_element_gap(1.0),
+          cw_inter_character_gap(3.0),
+          cw_inter_word_gap(7.0),
+          dfcw_intra_element_gap(kDefaultDfcwIntraElementGap),
+          dfcw_inter_character_gap(kDefaultDfcwInterCharacterGap),
+          dfcw_inter_word_gap(kDefaultDfcwInterWordGap),
+          cw_fade_shape("none"),
+          cw_fade_in_ms(0),
+          cw_fade_out_ms(0),
+          cw_fade_slice_ms(5),
+          schedule_start_minute(0),
+          schedule_start_second(5),
+          schedule_repeat_minutes(10),
+          mode(ModeType::WSPR),
+          wspr({}),
+          qrss({}),
+          fskcw({}),
+          dfcw({}),
+          use_ini(false),
+          ini_filename(""),
+          wspr_dial_freq_set({}),
+          wspr_frequency_entries({}),
+          frequency_estimate_good(false),
+          band_gpio({})
+    {
+    }
+
+    ArgParserConfig(const ArgParserConfig &other)
+        : ArgParserConfig()
+    {
+        *this = other;
+    }
+
+    ArgParserConfig &operator=(const ArgParserConfig &other)
+    {
+        if (this == &other)
+        {
+            return *this;
+        }
+
+        transmit = other.transmit;
+        enable_on_boot = other.enable_on_boot;
+        callsign = other.callsign;
+        grid_square = other.grid_square;
+        power_dbm = other.power_dbm;
+        frequencies = other.frequencies;
+        tx_pin = other.tx_pin;
+        ppm = other.ppm;
+        use_system_clock_frequency_estimate = other.use_system_clock_frequency_estimate;
+        use_offset = other.use_offset;
+        power_level = other.power_level;
+        transmit_backend = other.transmit_backend;
+        gpio_tx_pin = other.gpio_tx_pin;
+        gpio_power_level = other.gpio_power_level;
+        rp1_gpio_drive_ma = other.rp1_gpio_drive_ma;
+        gpio_use_system_clock_frequency_estimate = other.gpio_use_system_clock_frequency_estimate;
+        gpio_frequency_residual_ppm = other.gpio_frequency_residual_ppm;
+        gpio_manual_ppm = other.gpio_manual_ppm;
+        si5351_ppm = other.si5351_ppm;
+        si5351_i2c_bus = other.si5351_i2c_bus;
+        si5351_i2c_address = other.si5351_i2c_address;
+        si5351_reference_hz = other.si5351_reference_hz;
+        si5351_reference_source = other.si5351_reference_source;
+        si5351_crystal_load_capacitance_pf = other.si5351_crystal_load_capacitance_pf;
+        si5351_tx_output = other.si5351_tx_output;
+        si5351_power_level = other.si5351_power_level;
+        use_led = other.use_led;
+        led_pin = other.led_pin;
+        use_amp = other.use_amp;
+        amp_pin = other.amp_pin;
+        amp_pin_active_high = other.amp_pin_active_high;
+        enable_web = other.enable_web;
+        enable_http = other.enable_http;
+        web_port = other.web_port;
+        socket_port = other.socket_port;
+        socket_loopback_only = other.socket_loopback_only;
+        socket_loopback_family = other.socket_loopback_family;
+        use_shutdown = other.use_shutdown;
+        shutdown_pin = other.shutdown_pin;
+        use_journald = other.use_journald;
+        date_time_log = other.date_time_log;
+        debug_logging = other.debug_logging;
+        allow_unqualified_frequency = other.allow_unqualified_frequency;
+        allow_non_amateur_frequency = other.allow_non_amateur_frequency;
+        rp1_development_confirmation_json =
+            other.rp1_development_confirmation_json;
+        wspr_planner_preference = other.wspr_planner_preference;
+        loop_tx = other.loop_tx;
+        tx_iterations.store(other.tx_iterations.load());
+        wspr_audio_offset_hz = other.wspr_audio_offset_hz;
+        modulation_dot_seconds = other.modulation_dot_seconds;
+        modulation_fsk_offset_hz = other.modulation_fsk_offset_hz;
+        cw_intra_element_gap = other.cw_intra_element_gap;
+        cw_inter_character_gap = other.cw_inter_character_gap;
+        cw_inter_word_gap = other.cw_inter_word_gap;
+        dfcw_intra_element_gap = other.dfcw_intra_element_gap;
+        dfcw_inter_character_gap = other.dfcw_inter_character_gap;
+        dfcw_inter_word_gap = other.dfcw_inter_word_gap;
+        cw_fade_shape = other.cw_fade_shape;
+        cw_fade_in_ms = other.cw_fade_in_ms;
+        cw_fade_out_ms = other.cw_fade_out_ms;
+        cw_fade_slice_ms = other.cw_fade_slice_ms;
+        schedule_start_minute = other.schedule_start_minute;
+        schedule_start_second = other.schedule_start_second;
+        schedule_repeat_minutes = other.schedule_repeat_minutes;
+        mode = other.mode;
+        wspr = other.wspr;
+        qrss = other.qrss;
+        fskcw = other.fskcw;
+        dfcw = other.dfcw;
+        use_ini = other.use_ini;
+        ini_filename = other.ini_filename;
+        wspr_dial_freq_set = other.wspr_dial_freq_set;
+        wspr_frequency_entries = other.wspr_frequency_entries;
+        frequency_estimate_good = other.frequency_estimate_good;
+        band_gpio = other.band_gpio;
+        return *this;
+    }
+};
+
+/**
+ * @brief Global configuration object.
+ *
+ * This ArgParserConfig instance holds the application’s configuration settings,
+ * typically loaded from an INI file or a JSON configuration.
+ */
+extern ArgParserConfig config;
+
+/**
+ * @brief Immutable configuration inputs used to plan an explicit Test Tone.
+ *
+ * Published only after an accepted global configuration update.  Callers get
+ * a value copy so selector/frequency planning cannot observe mixed revisions.
+ */
+struct TestTonePlanningConfigSnapshot
+{
+    TransmitBackendKind transmit_backend = TransmitBackendKind::GPIO;
+    bool allow_unqualified_frequency = false;
+    bool allow_non_amateur_frequency = false;
+    double wspr_audio_offset_hz = WSPR_AUDIO_OFFSET_HZ;
+    std::string wspr_frequency_profile = "existing_common";
+    WsprBandPreferences wspr_band_preferences;
+    std::vector<WsprFrequencyEntry> wspr_frequency_entries{};
+    std::array<BandGPIOConfig, HAM_BAND_COUNT> band_gpio{};
+};
+
+#endif // CONFIG_TYPES_HPP
