+++ src/tests/ui_source_regression_test.cpp
@@ -93,7 +93,8 @@ namespace
 int main()
 {
     const std::string websocket_source =
-        read_repo_text_file("/src/web_socket.cpp");
+        read_repo_text_file("/src/web_socket.cpp") +
+        read_repo_text_file("/src/web_socket_commands.cpp");
     const std::string websocket_header_source =
         read_repo_text_file("/src/web_socket.hpp");
     const std::string web_server_source =
@@ -121,7 +122,9 @@ int main()
         "websocket Test Tone Start and End commands must share one transaction lock through their broadcast reply");
 
     const std::string scheduling_source =
-        read_repo_text_file("/src/scheduling.cpp");
+        read_repo_text_file("/src/scheduling.cpp") +
+        read_repo_text_file("/src/scheduling_config.cpp") +
+        read_repo_text_file("/src/scheduling_websocket_status.cpp");
     require(
         scheduling_source.find("suppress_cancelled_ws_event_for_user_stop.store(true") != std::string::npos &&
             scheduling_source.find("Suppressing websocket canceled event because an explicit user stop will publish stopped.") !=
