+++ src/tests/web_server_network_lifecycle_test.cpp
@@ -1,5 +1,6 @@
 #include "../privileged_network_runtime.hpp"
 #include "../web_server.hpp"
+#include "../httplib.hpp"
 
 #include <atomic>
 #include <cassert>
